requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
py-cord==2.6.0
py-dotenv==0.1
python-dotenv==1.0.1
yarl==1.9.4
